added auto type multiply and ratio.
the units_multiply and units_ratio function requires than user
to choose the result unit of the multiplication.

sometimes, when writing long calculations in physics and math,
you want to specify the result unit only in the end of the calculation.

for that, we created the auto_units_multiply and auto_units_ratio,
in which you can multiply and divide units values without specify the result type.

Pay attention: use this functions with care